Fix Visual Studio warnings in Vulkan backend

+++ Backends/Graphics5/Vulkan/Sources/Kore/Shader5Impl.cpp
@@ -3,7 +3,7 @@
 #include <kinc/graphics5/shader.h>
 
 void kinc_g5_shader_init(kinc_g5_shader_t *shader, void *source, size_t length, kinc_g5_shader_type_t type) {
-	shader->impl.length = length;
+	shader->impl.length = (int)length;
 	shader->impl.id = 0;
 	shader->impl.source = new char[length + 1];
 	for (int i = 0; i < length; ++i) {
